web_app/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.6
python-socketio>=5.10.0
eventlet>=0.33.3
pyserial>=3.5
pyyaml>=6.0
numpy>=1.24.0
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0